% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_acmg_pm2.R
\name{apply_acmg_pm2}
\alias{apply_acmg_pm2}
\title{Apply ACMG PM2 Criterion}
\usage{
apply_acmg_pm2(df, gnomad_max = 1e-06)
}
\arguments{
\item{df}{A dataframe containing genetic data.}

\item{gnomad_max}{The maximum allele frequency considered for PM2.}
}
\value{
A modified dataframe with the ACMG PM2 criterion applied.
}
\description{
Applies the ACMG PM2 criterion based on allele frequency in population databases.
}
