% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_acmg_pm3.R
\name{apply_acmg_pm3}
\alias{apply_acmg_pm3}
\title{Apply ACMG PM3 Criterion}
\usage{
apply_acmg_pm3(df)
}
\arguments{
\item{df}{A dataframe containing genetic data.}
}
\value{
A modified dataframe with the ACMG PM3 criterion applied.
}
\description{
Applies the ACMG PM3 criterion for recessive disorders detected in trans with a pathogenic variant.
}
