% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_acmg_pp3.R
\name{apply_acmg_pp3}
\alias{apply_acmg_pp3}
\title{Apply ACMG PP3 Criterion}
\usage{
apply_acmg_pp3(df)
}
\arguments{
\item{df}{A dataframe containing genetic data with necessary predictive columns.}
}
\value{
A modified dataframe with the ACMG PP3 criterion applied.
}
\description{
Applies the ACMG PP3 criterion based on multiple lines of computational evidence supporting a deleterious effect on the gene or gene product. This function assesses variants for evidence of pathogenicity based on computational predictions.
}
