% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_acmg_ps1.R
\name{apply_acmg_ps1}
\alias{apply_acmg_ps1}
\title{Apply ACMG PS1 Criterion}
\usage{
apply_acmg_ps1(df)
}
\arguments{
\item{df}{A dataframe containing genetic data with a \code{CLIN_SIG} column.}
}
\value{
A modified dataframe with the ACMG PS1 criterion applied.
}
\description{
This function applies the ACMG PS1 criterion, identifying pathogenic variants based on clinical significance.
}
