% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_acmg_ps5.R
\name{apply_acmg_ps5}
\alias{apply_acmg_ps5}
\title{Apply ACMG PS5 Criterion}
\usage{
apply_acmg_ps5(df)
}
\arguments{
\item{df}{A dataframe containing genetic data.}
}
\value{
A modified dataframe with the ACMG PS5 criterion applied.
}
\description{
Applies the ACMG PS5 criterion based on strong pathogenic evidence.
}
