% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_acmg_pvs1.R
\name{apply_acmg_pvs1}
\alias{apply_acmg_pvs1}
\title{Apply ACMG PVS1 Criterion}
\usage{
apply_acmg_pvs1(df)
}
\arguments{
\item{df}{A dataframe containing genetic data.}
}
\value{
A modified dataframe with the ACMG PVS1 criterion applied.
}
\description{
This function applies the ACMG PVS1 criterion to a given dataframe.
}
