% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_column_classes.R
\name{apply_column_classes}
\alias{apply_column_classes}
\title{Apply Column Classes Based on Reference Metadata}
\usage{
apply_column_classes(processed_data_list, metadataclass_file_path)
}
\arguments{
\item{processed_data_list}{A list of data frames to adjust.}

\item{metadataclass_file_path}{Path to the CSV file containing reference metadata for column classes.}
}
\value{
A list of adjusted data frames with updated column classes.
}
\description{
Adjusts the column classes of a list of data frames based on a provided reference metadata CSV file.
}
\examples{
processed_data_list <- list(df1 = data.frame(a = as.character(1:3)), df2 = data.frame(a = as.character(4:6)))
metadataclass_file_path <- system.file("extdata", "metadata.csv", package = "YourPackageName")
apply_column_classes(processed_data_list, metadataclass_file_path)
}
