% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_column_classes_to_processed_data.R
\name{apply_column_classes_to_processed_data}
\alias{apply_column_classes_to_processed_data}
\title{Apply Column Classes to Processed Data}
\usage{
apply_column_classes_to_processed_data(
  processed_data_list,
  metadataclass_file_path
)
}
\arguments{
\item{processed_data_list}{A list of dataframes to be adjusted.}

\item{metadataclass_file_path}{Path to the CSV file containing reference metadata for column classes.}
}
\description{
Applies the column classes to each dataframe in a list based on a provided reference metadata CSV file.
}
