% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_acmg_varsome_in_silico.R
\name{apply_varsome_in_silico}
\alias{apply_varsome_in_silico}
\title{Apply Varsome In-silico Predictions for ACMG PP3 Criterion}
\usage{
apply_varsome_in_silico(
  df,
  varsome_file_path = NULL,
  pathogenic_types = c("Strong_pathogenic_GE", "Moderate_pathogenic_GE",
    "Supporting_pathogenic_GE")
)
}
\arguments{
\item{df}{A dataframe containing genetic data with necessary predictive columns.}

\item{pathogenic_types}{A character vector of pathogenic types to assess ("Strong_pathogenic_GE",
"Moderate_pathogenic_GE", "Supporting_pathogenic_GE").}

\item{varsome_data}{A dataframe containing Varsome in-silico prediction scores and thresholds.}
}
\value{
A dataframe with additional columns indicating the count of pathogenic predictions met for each type.
}
\description{
Enhances the ACMG PP3 criterion by incorporating Varsome in-silico
prediction scores to assess variant pathogenicity.
}
