% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_acmg_varsome_in_silico.R
\name{calculate_varsome_score}
\alias{calculate_varsome_score}
\title{Calculate Varsome Score}
\usage{
calculate_varsome_score(df, varsome_data, pathogenic_type)
}
\description{
Internal function to calculate scores based on Varsome in-silico predictions.
}
