% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_summaries.R
\name{plot_criteria_count_each_gene}
\alias{plot_criteria_count_each_gene}
\title{Plotting Summaries for ACMG Criteria}
\usage{
plot_criteria_count_each_gene(df, file_suffix)
}
\description{
Contains functions for generating various plots to summarize the counts and distributions of ACMG criteria across genes and variants.
}
