% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_acmg_labels.R
\name{prepare_acmg_labels}
\alias{prepare_acmg_labels}
\title{Prepare ACMG Labels}
\usage{
prepare_acmg_labels(df)
}
\arguments{
\item{df}{A dataframe containing genetic data potentially including various ACMG label columns.}
}
\value{
A modified dataframe with additional columns for the highest priority ACMG classification (\code{ACMG_highest}) and the count of non-NA ACMG labels (\code{ACMG_count}) for each row.
}
\description{
Prepares ACMG labels by ensuring all expected ACMG label columns exist in the dataframe, assigning NA where they do not, identifying the highest priority ACMG classification for each row, and counting the number of non-NA ACMG labels for each row.
}
\examples{
df <- data.frame(
  ACMG_PVS1 = c(NA, "PVS1"),
  ACMG_PS1 = c("PS1", NA)
)
df_prepared <- prepare_acmg_labels(df)
}
