% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_genetic_data.R
\name{process_genetic_data}
\alias{process_genetic_data}
\title{Process Genetic Data According to ACMG Guidelines}
\usage{
process_genetic_data(input_path, samples_file_path, af_threshold)
}
\arguments{
\item{input_path}{Path to the input files or directory containing the CSV files.}

\item{samples_file_path}{Path to the samples file containing phenotype information.}

\item{af_threshold}{Allele frequency threshold for filtering variants.}
}
\value{
A list of data frames, each representing processed genetic data for a file.
}
\description{
This function processes genetic data by applying a series of ACMG criteria to identify variants based on specified thresholds and conditions.
}
\examples{
input_path <- system.file("extdata", package = "YourPackageName")
samples_file_path <- system.file("extdata", "samples.tsv", package = "YourPackageName")
af_threshold <- 0.01
processed_data_list <- process_genetic_data(input_path, samples_file_path, af_threshold)
}
