% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_file.R
\name{read_data_file}
\alias{read_data_file}
\title{Read Data Files and Merge with Sample Phenotype Information}
\usage{
read_data_file(input_path, samples_file_path)
}
\arguments{
\item{input_path}{Path to the input files or directory containing the files.}

\item{samples_file_path}{Path to the samples file containing phenotype information.}
}
\value{
A list of data frames, each representing a merged data file.
}
\description{
This function reads data files from a specified path and merges them with sample phenotype information.
}
\examples{
input_path <- system.file("extdata", package = "YourPackageName")
samples_file_path <- system.file("extdata", "samples.tsv", package = "YourPackageName")
read_data_file(input_path, samples_file_path)
}
