% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_file.R
\name{read_data_file}
\alias{read_data_file}
\title{Read Data Files and Merge with Sample Phenotype Information}
\usage{
read_data_file(input_path = NULL, samples_file_path, file_list = NULL)
}
\arguments{
\item{input_path}{Optional. Path to the input directory containing the files or a single file path. Used only if file_list is NULL.}

\item{samples_file_path}{Path to the samples file containing phenotype information.}

\item{file_list}{Optional. A vector of specific file paths to be processed. Overrides input_path if provided.}
}
\value{
A list of data frames, each representing a merged data file.
}
\description{
This function reads data files from a specified path and merges them with sample phenotype information. It can handle a single file, all files within a directory, or a specific list of files.
}
